requirements.txt
Django==5.2.3
djangorestframework==3.14.0
django-cors-headers==4.1.0
django-allauth==0.63.3
django-phonenumber-field==7.3.0
phonenumbers==8.13.40
python-dotenv==1.0.0
whitenoise==6.9.0
Pillow==11.3.0
opencv-python==4.12.0.88
pytesseract==0.3.13
tensorflow==2.20.0
numpy==2.2.6
requests==2.32.4
fuzzywuzzy==0.18.0
python-levenshtein==0.27.0